real_estate_bi/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-dateutil==2.8.2
psycopg2-binary==2.9.9